+++ app/models/bpsc_read_on_site_request.rb
@@ -5,6 +5,7 @@ class BPSCReadOnSiteRequest
                 :email,
                 :appointment_time,
                 :title,
+                :call_number,
                 :item_url,
                 :notes
 
@@ -12,5 +13,6 @@ class BPSCReadOnSiteRequest
   validates :email, presence: true
   validates :appointment_time, presence: true
   validates :title, presence: true
+  validates :call_number, presence: true
   validates :item_url, presence: true
 end